+++ project-biology/pb.bll/simulation-date.h
@@ -4,7 +4,7 @@
 
 void switchDate(unsigned short int&, std::string&);
 
-void timer(unsigned short int& day, std::string& month, bool isPaused, unsigned short int& choice);
+void timer(unsigned short int& day, std::string& month, bool isPaused, size_t& choice);
 
 int getIndex(std::vector<std::string>, std::string);
